wip: move functions to other files
