+++ web/src/network/trackUtils.ts
@@ -30,3 +30,40 @@ export const setupTrackStopOnLongMute = (
   track.addEventListener("mute", onmute);
   return track;
 };
+
+export const loopbackPeerConnection = (
+  track: MediaStreamTrack
+): Promise<MediaStreamTrack> =>
+  // eslint-disable-next-line no-async-promise-executor
+  new Promise(async (resolve, reject) => {
+    try {
+      const pcIn = new RTCPeerConnection();
+      const pcOut = new RTCPeerConnection();
+      pcIn.addEventListener("icecandidate", ({ candidate }) => {
+        if (candidate) {
+          pcOut.addIceCandidate(candidate);
+        }
+      });
+      pcOut.addEventListener("icecandidate", ({ candidate }) => {
+        if (candidate) {
+          pcIn.addIceCandidate(candidate);
+        }
+      });
+      pcOut.addEventListener("track", (event) => {
+        resolve(event.track);
+      });
+      track.addEventListener("ended", () => {
+        pcIn.close();
+        pcOut.close();
+      });
+      pcIn.addTrack(track);
+      const offer = await pcIn.createOffer();
+      await pcIn.setLocalDescription(offer);
+      await pcOut.setRemoteDescription(offer);
+      const answer = await pcOut.createAnswer();
+      await pcOut.setLocalDescription(answer);
+      await pcIn.setRemoteDescription(answer);
+    } catch (e) {
+      reject(e);
+    }
+  });
